Rename tests correctly
